requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
alembic
aioodbc
python-jose
passlib[bcrypt]
pydantic
email-validator
/* pip install -r requirements.txt */